pyproject.toml
[tool.poetry]
name = "fitness-house-bot"
version = "0.1.0"
description = ""
authors = ["Alexey Sergeev <alexey.nrdhm@gmail.com>"]
readme = "README.md"
packages = [{include = "fitness_house_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {version = "^20.0a6", allow-prereleases = true}
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
